Refactor code to create a sequence drawing module
