+++ sequence/colors.py
@@ -1,0 +1,8 @@
+import logging
+
+# Get logger object
+logger = logging.getLogger(__name__)
+
+class Colors:
+        black = '#000000'
+        gray = '#C0C0C0'
